backend/requirements.txt
Flask>=2.0.1
Flask-CORS>=3.0.10
requests>=2.0.0
python-dotenv>=0.19.0
numpy>=1.20.0
opencv-python>=4.5.3
pillow>=8.3.1
sympy>=1.8.0
matplotlib>=3.4.3
pytesseract>=0.3.8
python-dotenv>=1.0.0